new icon set, button styles
It's not great, but it's better than what was there. Which was the old paint can. And I'm not sure anyone knew what it was who didn't know it already. This is plain, but it works. 'Till we get something better.

Also added the button style and changed the overlay ID to match our plugin instead of the one we stole it from.